requirements.txt
streamlit
langchain
langchain-groq
langchain-google-genai
langchain-community
textblob
tavily-python
python-dotenv